Diagrams.Backend.Cairo.CmdLine

Description

Convenient creation of command-line-driven executables for rendering diagrams using the cairo backend.

Documentation

defaultMain :: Diagram Cairo R2 -> IO ()

This is the simplest way to render diagrams, and is intended to be used like so:

 ... definitions ...

 main = defaultMain myDiagram

Compiling this file will result in an executable which takes various command-line options for setting the size, output file, and so on, and renders myDiagram with the specified options.

On Unix systems, the generated executable also supports a rudimentary "looped" mode, which watches the source file for changes and recompiles itself on the fly.

Pass --help to the generated executable to see all available options.

multiMain :: [(String, Diagram Cairo R2)] -> IO ()

multiMain is like defaultMain, except instead of a single diagram it takes a list of diagrams paired with names as input. The generated executable then takes an argument specifying the name of the diagram that should be rendered. This is a convenient way to create an executable that can render many different diagrams without modifying the source code in between each one.

animMain :: Animation Cairo R2 -> IO ()

XXX

data Cairo

This data declaration is simply used as a token to distinguish this rendering engine.