Diagrams.Backend.Cairo.Internal

Description

Implementation of the diagrams Cairo backend. This module exports the internal helper functions, for usage elsewhere.

Documentation

data Cairo

This data declaration is simply used as a token to distinguish this rendering engine.

Constructors

Cairo

data OutputType

Output types supported by cairo, including four different file types (PNG, PS, PDF, SVG) as well as Gtk windows.

Constructors

forall dw . DrawableClass dw => GTK
Output directly to a Gtk window. See Diagrams.Backends.Cairo.Gtk.

Fields

gtkWindow :: dw

The window on which to draw.

gtkBypass :: Bool

Should the adjustDia step be bypassed during rendering?

PNG
Portable Network Graphics output.

PS
PostScript output

PDF
Portable Document Format output.

SVG
Scalable Vector Graphics output.

type RenderM a = StateT () Render a

save :: RenderM ()

restore :: RenderM ()

renderC :: (Renderable a Cairo, V a ~ R2) => a -> RenderM ()

cairoMiscStyle :: Style v -> RenderM ()

fromFontSlant :: FontSlant -> FontSlant

fromFontWeight :: FontWeight -> FontWeight

cairoStrokeStyle :: Style v -> Render ()

setSource :: Color c => c -> Style v -> Render ()

cairoTransf :: T2 -> Render ()

fromLineCap :: LineCap -> LineCap

fromLineJoin :: LineJoin -> LineJoin

fromFillRule :: FillRule -> FillRule